Aforro_Intern_Task/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
celery
redis
django-redis       
django-filter
drf-spectacular
gunicorn